+++ src/dosbox.cpp
@@ -229,14 +229,14 @@ void increaseticks() { //Make it return ticksRemain and set it in the function a
 		Bit32s ratio = (ticksScheduled * (CPU_CyclePercUsed*90*1024/100/100)) / ticksDone;
 		Bit32s new_cmax = CPU_CycleMax;
 		Bit64s cproc = (Bit64s)CPU_CycleMax * (Bit64s)ticksScheduled;
-		double ratioremoved = 0.0; //increase scope for logging
+		float ratioremoved = 0.0; //increase scope for logging
 		if (cproc > 0) {
 			/* ignore the cycles added due to the IO delay code in order
 			   to have smoother auto cycle adjustments */
-			ratioremoved = (double) CPU_IODelayRemoved / (double) cproc;
+			ratioremoved = (float) CPU_IODelayRemoved / (float) cproc;
 			if (ratioremoved < 1.0) {
-				double ratio_not_removed = 1 - ratioremoved;
-				ratio = (Bit32s)((double)ratio * ratio_not_removed);
+				float ratio_not_removed = 1 - ratioremoved;
+				ratio = (Bit32s)((float)ratio * ratio_not_removed);
 
 				/* Don't allow very high ratio which can cause us to lock as we don't scale down
 				 * for very low ratios. High ratio might result because of timing resolution */
@@ -253,10 +253,10 @@ void increaseticks() { //Make it return ticksRemain and set it in the function a
 
 				if (ratio <= 1024) {
 					// ratio_not_removed = 1.0; //enabling this restores the old formula
-					double r = (1.0 + ratio_not_removed) /(ratio_not_removed + 1024.0/(static_cast<double>(ratio)));
+					float r = (1.0 + ratio_not_removed) /(ratio_not_removed + 1024.0/(static_cast<float>(ratio)));
 					new_cmax = 1 + static_cast<Bit32s>(CPU_CycleMax * r);
 				} else {
-					Bit64s ratio_with_removed = (Bit64s) ((((double)ratio - 1024.0) * ratio_not_removed) + 1024.0);
+					Bit64s ratio_with_removed = (Bit64s) ((((float)ratio - 1024.0) * ratio_not_removed) + 1024.0);
 					Bit64s cmax_scaled = (Bit64s)CPU_CycleMax * ratio_with_removed;
 					new_cmax = (Bit32s)(1 + (CPU_CycleMax >> 1) + cmax_scaled / (Bit64s)2048);
 				}
