+++ src/hardware/timer.cpp
@@ -26,6 +26,7 @@
 #include "timer.h"
 #include "setup.h"
 
+
 static INLINE void BIN2BCD(Bit16u& val) {
 	Bit16u temp=val%10 + (((val/10)%10)<<4)+ (((val/100)%10)<<8) + (((val/1000)%10)<<12);
 	val=temp;
@@ -39,7 +40,7 @@ static INLINE void BCD2BIN(Bit16u& val) {
 struct PIT_Block {
 	Bitu cntr;
 	float delay;
-	double start;
+	float start;
 
 	Bit16u read_latch;
 	Bit16u write_latch;
@@ -80,7 +81,7 @@ static void PIT0_Event(Bitu /*val*/) {
 
 static bool counter_output(Bitu counter) {
 	PIT_Block * p=&pit[counter];
-	double index=PIC_FullIndex()-p->start;
+	float index=PIC_FullIndex()-p->start;
 	switch (p->mode) {
 	case 0:
 		if (p->new_mode) return false;
@@ -89,15 +90,15 @@ static bool counter_output(Bitu counter) {
 		break;
 	case 2:
 		if (p->new_mode) return true;
-		index=fmod(index,(double)p->delay);
+		index=fmodf(index,(float)p->delay);
 		return index>0;
 	case 3:
 		if (p->new_mode) return true;
-		index=fmod(index,(double)p->delay);
+		index=fmodf(index,(float)p->delay);
 		return index*2<p->delay;
 	case 4:
 		//Only low on terminal count
-		// if(fmod(index,(double)p->delay) == 0) return false; //Maybe take one rate tick in consideration
+		// if(fmod(index,(float)p->delay) == 0) return false; //Maybe take one rate tick in consideration
 		//Easiest solution is to report always high (Space marines uses this mode)
 		return true;
 	default:
@@ -139,13 +140,13 @@ static void counter_latch(Bitu counter) {
 	//If gate2 is disabled don't update the read_latch
 	if (counter == 2 && !gate2 && p->mode !=1) return;
 	if (GCC_UNLIKELY(p->new_mode)) {
-		double passed_time = PIC_FullIndex() - p->start;
+		float passed_time = PIC_FullIndex() - p->start;
 		Bitu ticks_since_then = (Bitu)(passed_time / (1000.0/PIT_TICK_RATE));
 		//if (p->mode==3) ticks_since_then /= 2; // TODO figure this out on real hardware
 		p->read_latch -= ticks_since_then;
 		return;
 	}
-	double index=PIC_FullIndex()-p->start;
+	float index=PIC_FullIndex()-p->start;
 	switch (p->mode) {
 	case 4:		/* Software Triggered Strobe */
 	case 0:		/* Interrupt on Terminal Count */
@@ -153,10 +154,10 @@ static void counter_latch(Bitu counter) {
 		if (index>p->delay) {
 			index-=p->delay;
 			if(p->bcd) {
-				index = fmod(index,(1000.0/PIT_TICK_RATE)*10000.0);
+				index = fmodf(index,(1000.0/PIT_TICK_RATE)*10000.0);
 				p->read_latch = (Bit16u)(9999-index*(PIT_TICK_RATE/1000.0));
 			} else {
-				index = fmod(index,(1000.0/PIT_TICK_RATE)*(double)0x10000);
+				index = fmodf(index,(1000.0/PIT_TICK_RATE)*(float)0x10000);
 				p->read_latch = (Bit16u)(0xffff-index*(PIT_TICK_RATE/1000.0));
 			}
 		} else {
@@ -173,11 +174,11 @@ static void counter_latch(Bitu counter) {
 		}
 		break;
 	case 2:		/* Rate Generator */
-		index=fmod(index,(double)p->delay);
+		index=fmodf(index,(float)p->delay);
 		p->read_latch=(Bit16u)(p->cntr - (index/p->delay)*p->cntr);
 		break;
 	case 3:		/* Square Wave Rate Generator */
-		index=fmod(index,(double)p->delay);
+		index=fmodf(index,(float)p->delay);
 		index*=2;
 		if (index>p->delay) index-=p->delay;
 		p->read_latch=(Bit16u)(p->cntr - (index/p->delay)*p->cntr);
@@ -242,7 +243,7 @@ static void write_latch(Bitu port,Bitu val,Bitu /*iolen*/) {
 			LOG(LOG_PIT,LOG_NORMAL)("PIT 0 Timer at %.4f Hz mode %d",1000.0/p->delay,p->mode);
 			break;
 		case 0x02:			/* Timer hooked to PC-Speaker */
-//			LOG(LOG_PIT,"PIT 2 Timer at %.3g Hz mode %d",PIT_TICK_RATE/(double)p->cntr,p->mode);
+//			LOG(LOG_PIT,"PIT 2 Timer at %.3g Hz mode %d",PIT_TICK_RATE/(float)p->cntr,p->mode);
 			PCSPEAKER_SetCounter(p->cntr,p->mode);
 			break;
 		default:
